cloud/backend/cloud_tests/monitoring/requirements.txt
requests==2.14.2
boto3==1.4.4
docker==3.5.0
pyOpenSSL==19.0.0
